requirements.txt
gradio
google-genai
smolagents
duckduckgo-search